requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file ./requirements.txt ./requirements.in
#
asn1crypto==0.24.0        # via cryptography
attrs==17.4.0             # via automat, service-identity
automat==0.6.0            # via twisted
certifi==2017.11.5        # via requests
cffi==1.11.2              # via cryptography
chardet==3.0.4            # via requests
constantly==15.1.0        # via twisted
cryptography==2.1.4       # via pyopenssl
cssselect==1.0.3          # via parsel, scrapy
django-cors-headers==2.1.0
django-rest-framework==0.1.0
django==2.0.1
djangorestframework==3.7.7  # via django-rest-framework
hyperlink==17.3.1         # via twisted
idna==2.6                 # via cryptography, requests
incremental==17.5.0       # via twisted
lxml==4.1.1               # via parsel, scrapy
parsel==1.3.1             # via scrapy
pyasn1-modules==0.2.1     # via service-identity
pyasn1==0.4.2             # via pyasn1-modules, service-identity
pycparser==2.18           # via cffi
pydispatcher==2.0.5       # via scrapy
pyopenssl==17.5.0         # via scrapy, service-identity
pytz==2017.3              # via django
queuelib==1.4.2           # via scrapy
requests==2.18.4
scrapy-djangoitem==1.1.1
scrapy==1.5.0
service-identity==17.0.0  # via scrapy
six==1.11.0               # via automat, cryptography, parsel, pyopenssl, scrapy, scrapy-djangoitem, w3lib
twisted==17.9.0           # via scrapy
urllib3==1.22             # via requests
w3lib==1.18.0             # via parsel, scrapy
zope.interface==4.4.3     # via twisted
